Refactor weekly summary tooling into package
