+++ tools/__init__.py
@@ -1,0 +1,1 @@
+"""Utility scripts for repository maintenance."""
