+++ tools/weekly_summary/__init__.py
@@ -1,0 +1,48 @@
+from __future__ import annotations
+
+from .data import (
+    count_new_defects,
+    extract_defect_dates,
+    filter_by_window,
+    load_flaky,
+    load_runs,
+    parse_iso8601,
+    select_flaky_rows,
+)
+from .io import fallback_write, write_summary
+from .report import (
+    SummaryData,
+    aggregate_status,
+    build_front_matter,
+    build_markdown,
+    compute_failure_top,
+    compute_summary,
+    ensure_front_matter,
+    format_percentage,
+    format_table,
+    to_float,
+    week_over_week_notes,
+)
+
+__all__ = [
+    "SummaryData",
+    "aggregate_status",
+    "build_front_matter",
+    "build_markdown",
+    "compute_failure_top",
+    "compute_summary",
+    "count_new_defects",
+    "extract_defect_dates",
+    "fallback_write",
+    "filter_by_window",
+    "format_percentage",
+    "format_table",
+    "load_flaky",
+    "load_runs",
+    "parse_iso8601",
+    "select_flaky_rows",
+    "to_float",
+    "week_over_week_notes",
+    "write_summary",
+    "ensure_front_matter",
+]
